Change1 to README.md
Dear Reader, 

This is the first of many changes to my read.md file, thank you for reading this description and following me along on my software developer journey. If you are starting out, I wish you the best. If you've been coding a while, give me a follow so that I can learn from you and your work and help to contribute code for future developers.